requirements.txt
blinker==1.7.0
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
Flask==3.0.2
functions-framework==3.5.0
google-cloud==0.34.0
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
watchdog==4.0.0
Werkzeug==3.0.1
zipp==3.17.0
google-cloud-storage>=1.42.0